requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
decouple==0.0.7
ecdsa==0.18.0
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna==3.4
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.10.4
PyMySQL==1.0.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
